UX improvements
Should work with larger files (probably), with third party cookies blocked, metadata format should follow FFprobe's and UI should be a bit better now.

Also increase rants about the File System API: it kinda remembers me new ActiveXObject("Microsoft.XMLHTTP"): a bad solution that nobody now uses but at that time was the only one for that problem and the one that people had to use. Good thing that Microsoft created it, but we moved from that! When will that new FS Standard get forward? No pressure but... yeah.